Make use_compat a public flag to array_namespace()
If None, it does the default behavior. If True or False it forces the use or
non-use of the compat wrapper.

This commit also changes it so that NumPy 2.0 does not return the wrapper
library at all by default (when use_compat=None), since it is completely array
API compatible on its own.